CIP-0015: erased the signature key 2
It was decided to reuse key 1, but change the signature algorithm to the
one that I aimed to introduce in an unambiguous, backward compatible
way under key 2.

Future examiners of registration metadata on the blockchain should be
advised to use secondary data to distinguish the different formats: the
new format registrations should carry the nonce under 61284 entry key 4
and are introduced from fund 4 onwards.