Update CIP-0027.md
This is an extensive rewrite of proposal CIP-0027.  All previous iterations are to be considered deprecated.  This proposal has been developed through extensive community outreach, and represents community compromise for scalability and simplicity.  The community intends to proceed with using this standard forthwith.  We respectfully ask that the CIP be considered a priority measure, given the significant impact it will have on the Cardano Community.