CIP-0027 | Standardization of royalties (#116)
* Adding CIP-0026 Standardization of royalties

* Renamed to CIP-0027 and added metadata example

* Update CIP-0027.md

Adjust Formatting

* Update CIP-0027.md

This is an extensive rewrite of proposal CIP-0027.  All previous iterations are to be considered deprecated.  This proposal has been developed through extensive community outreach, and represents community compromise for scalability and simplicity.  The community intends to proceed with using this standard forthwith.  We respectfully ask that the CIP be considered a priority measure, given the significant impact it will have on the Cardano Community.

* Update CIP-0027.md

Updated contributor section

* Update CIP-0027.md

Add line 23 to state clearly that the royalty token needs to be made prior to the assets under it.

* Update CIP-0027.md

Added clarification of the necessity to use a brand new policy for implementation

* Update and rename CIP-0027.md to CIP-0027/CIP-0027.md

Added header, and moved file under its own folder.

* Update registry.json

Adding CIP-0010 Registry update to CIP-0027 Pull request

* Update registry.json

Fixed typo

* Update CIP-0027.md

Have replaced pct tag with rate, and notated that legacy pct tag to continue to be honored by marketplaces.

Co-authored-by: John <john@ADA-Relay2.beyondvoip.net>