Update and rename Decentralization - Using Pledge as a Bidding Parameter to Determine Saturation Limit.md to CIP Decentralization: Using Pledge as a Bidding Parameter to Determine Saturation Limit/README.md
Hi @rphair,
I revised the proposal in accordance with your suggestions. It took a while since I needed to get used to the GitHub workflow, but I think I may be getting better at it now. Please let me know if there is anything else that needs to be done, and thank you for your help!

Updates:
1. File was moved to a subdirectory prefixed with "CIP".
2. File was renamed to README.md.
3. The header was wrapped with --- and should now appear in a table.
4. The header now includes Comments-URI.
5. File now contains standard Markdown syntax (line breaks are introduced using <br> instead of \ and &nbsp;).
6. Typos were fixed and some necessary edits were made to further improve the proposal.