removed 'CBOR' from byte array
removed the 'CBOR' from the byte array in the example, because its not a CBOR byte array. if it would be a CBOR byte array we would need to have the CBOR major type 4 in it with the length of the byte array. which is not the case. its simply a byte array of the public_keys/rewards_address.