CIP-0025: Update "image" Spec Based on Feedback
Practical usage has led to a proliferation of URI mechanisms for the
"image" field in CIP-0025 metadata.  (See
https://twitter.com/pool_pm/status/1575219209491337216?s=20&t=KZKbGeCKIUnPyTZ0fzPyhQ
for one thread on the topic).  This change would be a simple
clarification in one sub-bullet for newly introduced community members
looking to get familiar with the standard.

Since IPFS and Arweave are the most common methods of data storage
for NFT files, we explicitly call them out and provide more resources.
Note that this is particularly important as certain free storage
providers like NFT.Storage only return the version 1 CID in their file
upload calls and lead to scheme being commonly omitted in the URI.