Draft CPS-004 on Voltaire era Governance
This CPS tried to frame a set of motivations, goals, constraints, and open questions for the very broad topic of Cardano governance.

Whereas discussions within a specific CIP are most effective when targeted at the specific technical solution in question, a CPS allows for broader philosophical discussion that may shape whether an entirely different CIP is needed.

Despite being generally in support of CIP-1694, I tried to keep this very neutral.  I reached out to several people to help review this, including a few that are mildly or staunchly opposed to CIP-1694, and their feedback was that I had been successful in that neutrality.

I also have an open offer: if anyone wants to write a competing CIP to 1694, I'm happy to help refine the technical argument to be the strongest proposal it can be.

I think ultimately the community will be strengthened by these perspectices, and even if the community does align behind 1694, other CIPs can provide useful insights to refining CIP-1694, or provide follow up work to continue improving governance beyond the initial minimum viable governance.