CIP-1694? | A First Step Towards On-Chain Decentralized Governance (#380)
* A proposal for entering the Voltaire phase

We're heading into the age of Voltaire, laying down the foundations for
decentralized decision-making. This CIP describes a mechanism for
on-chain governance that will underpin the Voltaire phase of Cardano.
The document builds on and extends the original Cardano governance
scheme that was based on a fixed number of governance keys.
It aims to provide a first step that is both valuable and technically
achievable in the near term as part of the proposed Voltaire governance
system.

It also seeks to act as a jumping-off point for continuing community
input, including on appropriate threshold settings and other on-chain
settings. Subsequent proposals may adapt and extend this proposal to
meet emerging governance needs.

* updated table modification times after checking merge conflict resolved OK

* orphaned or duplicated comment in top level README

If referring to CPS statuses, CIP-0001 link is incorrect & should be fixed in a separate PR (keeping this change unique to fixing this PR's conflict without having to rebase)

---------

Co-authored-by: Robert Phair <rphair@cosd.com>