Revision based on community feedback
This implements the feedback from 2 separate community workshops, the
recordings of which can be found here:

 - https://www.youtube.com/watch?v=X3XdQZx_TyU
 - https://www.youtube.com/watch?v=bBQMF6Dr6HY

In particular, this limits the scope of what this CIP tries to achieve,
and answers many of the open questions. There are still a few small
items to finish off before we're ready to merge this.