CIP-0005 | Distinguish between key based and script based conway credentials (#737)
* Update CIP05 to distinguish between key based and script based DREPs and CC members

Committee members and DReps can be identified by a credential that can be either an Ed25519 verification key OR  a Script. The current prefixes do not make any distinction between these. It is good for clarity and transparency to make a clear distinction between the two types.

* Update CIP105 drep prefix

Update to reflect drep prefix being for keybased dreps

* Updating cc credentials meaning on CIP105

* move credentials to hash subsection + align descriptions

* align changes with CIP105

* add clarification on credential naming

* tidy cip105

* added missing category
* fixed links to point to Intersect repository
* improved the path to active section

* add script bech32 prefixes to cip105

---------

Co-authored-by: Ryan Williams <44342099+Ryun1@users.noreply.github.com>