Create CIP Document
A key point of "[Social Governance](https://github.com/cardano-foundation/CIPs/pull/935)" that we have not been able to adequately represent in the Cardano Constitution is how DReps should establish a Code of Conduct to guide their decisions.  While such a code would be difficult to mandate in a permissionless blockchain system that lacks adequate enforcement mechanisms, we still feel it is important to offer a template that DReps and other governance actors can use to voluntarily create a personal code of conduct for themselves.

This CIP sets forth the recommended disclosures and process in hopes that most governance actors will choose to engage so that potential delegators can use this information to make a decision on who to support.